+++ cases/models/cylinder.step
@@ -1,0 +1,148 @@
+ISO-10303-21;
+HEADER;
+FILE_DESCRIPTION(('Open CASCADE Model'),'2;1');
+FILE_NAME('Open CASCADE Shape Model','2026-05-13T16:46:25',('Author'),(
+    'Open CASCADE'),'Open CASCADE STEP processor 7.7','Open CASCADE 7.7'
+  ,'Unknown');
+FILE_SCHEMA(('AUTOMOTIVE_DESIGN { 1 0 10303 214 1 1 1 1 }'));
+ENDSEC;
+DATA;
+#1 = APPLICATION_PROTOCOL_DEFINITION('international standard',
+  'automotive_design',2000,#2);
+#2 = APPLICATION_CONTEXT(
+  'core data for automotive mechanical design processes');
+#3 = SHAPE_DEFINITION_REPRESENTATION(#4,#10);
+#4 = PRODUCT_DEFINITION_SHAPE('','',#5);
+#5 = PRODUCT_DEFINITION('design','',#6,#9);
+#6 = PRODUCT_DEFINITION_FORMATION('','',#7);
+#7 = PRODUCT('Open CASCADE STEP translator 7.7 10',
+  'Open CASCADE STEP translator 7.7 10','',(#8));
+#8 = PRODUCT_CONTEXT('',#2,'mechanical');
+#9 = PRODUCT_DEFINITION_CONTEXT('part definition',#2,'design');
+#10 = ADVANCED_BREP_SHAPE_REPRESENTATION('',(#11,#15),#113);
+#11 = AXIS2_PLACEMENT_3D('',#12,#13,#14);
+#12 = CARTESIAN_POINT('',(0.,0.,0.));
+#13 = DIRECTION('',(0.,0.,1.));
+#14 = DIRECTION('',(1.,0.,-0.));
+#15 = MANIFOLD_SOLID_BREP('',#16);
+#16 = CLOSED_SHELL('',(#17,#105,#109));
+#17 = ADVANCED_FACE('',(#18),#32,.T.);
+#18 = FACE_BOUND('',#19,.F.);
+#19 = EDGE_LOOP('',(#20,#50,#77,#78));
+#20 = ORIENTED_EDGE('',*,*,#21,.T.);
+#21 = EDGE_CURVE('',#22,#24,#26,.T.);
+#22 = VERTEX_POINT('',#23);
+#23 = CARTESIAN_POINT('',(0.,15.,0.));
+#24 = VERTEX_POINT('',#25);
+#25 = CARTESIAN_POINT('',(50.,15.,0.));
+#26 = SEAM_CURVE('',#27,(#31,#43),.PCURVE_S1.);
+#27 = LINE('',#28,#29);
+#28 = CARTESIAN_POINT('',(0.,15.,0.));
+#29 = VECTOR('',#30,1.);
+#30 = DIRECTION('',(1.,0.,0.));
+#31 = PCURVE('',#32,#37);
+#32 = CYLINDRICAL_SURFACE('',#33,15.);
+#33 = AXIS2_PLACEMENT_3D('',#34,#35,#36);
+#34 = CARTESIAN_POINT('',(0.,0.,0.));
+#35 = DIRECTION('',(-1.,-0.,-0.));
+#36 = DIRECTION('',(0.,1.,0.));
+#37 = DEFINITIONAL_REPRESENTATION('',(#38),#42);
+#38 = LINE('',#39,#40);
+#39 = CARTESIAN_POINT('',(-0.,0.));
+#40 = VECTOR('',#41,1.);
+#41 = DIRECTION('',(-0.,-1.));
+#42 = ( GEOMETRIC_REPRESENTATION_CONTEXT(2) 
+PARAMETRIC_REPRESENTATION_CONTEXT() REPRESENTATION_CONTEXT('2D SPACE',''
+  ) );
+#43 = PCURVE('',#32,#44);
+#44 = DEFINITIONAL_REPRESENTATION('',(#45),#49);
+#45 = LINE('',#46,#47);
+#46 = CARTESIAN_POINT('',(-6.28318530718,0.));
+#47 = VECTOR('',#48,1.);
+#48 = DIRECTION('',(-0.,-1.));
+#49 = ( GEOMETRIC_REPRESENTATION_CONTEXT(2) 
+PARAMETRIC_REPRESENTATION_CONTEXT() REPRESENTATION_CONTEXT('2D SPACE',''
+  ) );
+#50 = ORIENTED_EDGE('',*,*,#51,.T.);
+#51 = EDGE_CURVE('',#24,#24,#52,.T.);
+#52 = SURFACE_CURVE('',#53,(#58,#65),.PCURVE_S1.);
+#53 = CIRCLE('',#54,15.);
+#54 = AXIS2_PLACEMENT_3D('',#55,#56,#57);
+#55 = CARTESIAN_POINT('',(50.,0.,0.));
+#56 = DIRECTION('',(1.,0.,0.));
+#57 = DIRECTION('',(0.,1.,0.));
+#58 = PCURVE('',#32,#59);
+#59 = DEFINITIONAL_REPRESENTATION('',(#60),#64);
+#60 = LINE('',#61,#62);
+#61 = CARTESIAN_POINT('',(-0.,-50.));
+#62 = VECTOR('',#63,1.);
+#63 = DIRECTION('',(-1.,0.));
+#64 = ( GEOMETRIC_REPRESENTATION_CONTEXT(2) 
+PARAMETRIC_REPRESENTATION_CONTEXT() REPRESENTATION_CONTEXT('2D SPACE',''
+  ) );
+#65 = PCURVE('',#66,#71);
+#66 = PLANE('',#67);
+#67 = AXIS2_PLACEMENT_3D('',#68,#69,#70);
+#68 = CARTESIAN_POINT('',(50.,-1.093952106032E-15,1.169780044473E-16));
+#69 = DIRECTION('',(1.,0.,0.));
+#70 = DIRECTION('',(-0.,0.,1.));
+#71 = DEFINITIONAL_REPRESENTATION('',(#72),#76);
+#72 = CIRCLE('',#73,15.);
+#73 = AXIS2_PLACEMENT_2D('',#74,#75);
+#74 = CARTESIAN_POINT('',(-1.169780044473E-16,-1.093952106032E-15));
+#75 = DIRECTION('',(0.,-1.));
+#76 = ( GEOMETRIC_REPRESENTATION_CONTEXT(2) 
+PARAMETRIC_REPRESENTATION_CONTEXT() REPRESENTATION_CONTEXT('2D SPACE',''
+  ) );
+#77 = ORIENTED_EDGE('',*,*,#21,.F.);
+#78 = ORIENTED_EDGE('',*,*,#79,.F.);
+#79 = EDGE_CURVE('',#22,#22,#80,.T.);
+#80 = SURFACE_CURVE('',#81,(#86,#93),.PCURVE_S1.);
+#81 = CIRCLE('',#82,15.);
+#82 = AXIS2_PLACEMENT_3D('',#83,#84,#85);
+#83 = CARTESIAN_POINT('',(0.,0.,0.));
+#84 = DIRECTION('',(1.,0.,0.));
+#85 = DIRECTION('',(0.,1.,0.));
+#86 = PCURVE('',#32,#87);
+#87 = DEFINITIONAL_REPRESENTATION('',(#88),#92);
+#88 = LINE('',#89,#90);
+#89 = CARTESIAN_POINT('',(-0.,0.));
+#90 = VECTOR('',#91,1.);
+#91 = DIRECTION('',(-1.,0.));
+#92 = ( GEOMETRIC_REPRESENTATION_CONTEXT(2) 
+PARAMETRIC_REPRESENTATION_CONTEXT() REPRESENTATION_CONTEXT('2D SPACE',''
+  ) );
+#93 = PCURVE('',#94,#99);
+#94 = PLANE('',#95);
+#95 = AXIS2_PLACEMENT_3D('',#96,#97,#98);
+#96 = CARTESIAN_POINT('',(0.,-1.093952106032E-15,1.169780044473E-16));
+#97 = DIRECTION('',(1.,0.,0.));
+#98 = DIRECTION('',(-0.,0.,1.));
+#99 = DEFINITIONAL_REPRESENTATION('',(#100),#104);
+#100 = CIRCLE('',#101,15.);
+#101 = AXIS2_PLACEMENT_2D('',#102,#103);
+#102 = CARTESIAN_POINT('',(-1.169780044473E-16,-1.093952106032E-15));
+#103 = DIRECTION('',(0.,-1.));
+#104 = ( GEOMETRIC_REPRESENTATION_CONTEXT(2) 
+PARAMETRIC_REPRESENTATION_CONTEXT() REPRESENTATION_CONTEXT('2D SPACE',''
+  ) );
+#105 = ADVANCED_FACE('',(#106),#94,.F.);
+#106 = FACE_BOUND('',#107,.F.);
+#107 = EDGE_LOOP('',(#108));
+#108 = ORIENTED_EDGE('',*,*,#79,.T.);
+#109 = ADVANCED_FACE('',(#110),#66,.T.);
+#110 = FACE_BOUND('',#111,.T.);
+#111 = EDGE_LOOP('',(#112));
+#112 = ORIENTED_EDGE('',*,*,#51,.T.);
+#113 = ( GEOMETRIC_REPRESENTATION_CONTEXT(3) 
+GLOBAL_UNCERTAINTY_ASSIGNED_CONTEXT((#117)) GLOBAL_UNIT_ASSIGNED_CONTEXT
+((#114,#115,#116)) REPRESENTATION_CONTEXT('Context #1',
+  '3D Context with UNIT and UNCERTAINTY') );
+#114 = ( LENGTH_UNIT() NAMED_UNIT(*) SI_UNIT(.MILLI.,.METRE.) );
+#115 = ( NAMED_UNIT(*) PLANE_ANGLE_UNIT() SI_UNIT($,.RADIAN.) );
+#116 = ( NAMED_UNIT(*) SI_UNIT($,.STERADIAN.) SOLID_ANGLE_UNIT() );
+#117 = UNCERTAINTY_MEASURE_WITH_UNIT(LENGTH_MEASURE(1.E-07),#114,
+  'distance_accuracy_value','confusion accuracy');
+#118 = PRODUCT_RELATED_PRODUCT_CATEGORY('part',$,(#7));
+ENDSEC;
+END-ISO-10303-21;
